Split user service
