+++ src/services/index.js
@@ -1,4 +1,6 @@
 export { default as userService } from './user';
+export { default as authService } from './auth';
+export { default as identityModeService } from './identityMode';
 export { default as tokenInfoService } from './tokeninfo';
 export { default as ethplorerService } from './ethplorer';
 export { default as priceService } from './price';
