+++ test/unit/mocks/services/auth.js
@@ -1,0 +1,11 @@
+import authService from '@/services/auth';
+
+jest.mock('@/services/auth', () => ({
+  login: jest.fn().mockResolvedValue('email_auth'),
+
+  logout: jest.fn().mockResolvedValue({ success: true }),
+
+  loginViaOTP: jest.fn().mockResolvedValue({ success: true }),
+}));
+
+export default authService;
