+++ test/unit/mocks/services/identityMode.js
@@ -1,0 +1,11 @@
+import identityModeService from '@/services/identityMode';
+
+jest.mock('@/services/identityMode', () => ({
+  setIdentityMode: jest.fn().mockReturnValue(),
+
+  getIdentityMode: jest.fn().mockReturnValue({ type: 'default' }),
+
+  deleteIdentityData: jest.fn().mockResolvedValue(),
+}));
+
+export default identityModeService;
