+++ test/unit/specs/services/user.spec.js
@@ -1,11 +1,7 @@
-import axios from 'axios';
-import { proxyRequest } from '@/utils';
 import MockAdapter from 'axios-mock-adapter';
 
 import { NotificationError } from '@/class';
 import { httpIdentity } from '@/class/singleton';
-import { IDENTITY_MODE } from '@/constants';
-import { addresses } from 'fixtures/accounts';
 
 const userService = require.requireActual('@/services/user').default;
 
@@ -14,151 +10,6 @@ describe('User service', () => {
 
   beforeEach(() => {
     axiosMock = new MockAdapter(httpIdentity);
-  });
-
-  describe('login', () => {
-    const redirectUri = '/send?to=0x1234&amount=0.1';
-    const url = `${
-      ENV.identityAPIUrl
-    }/auth?redirect_uri=%2Fsend%3Fto%3D0x1234%26amount%3D0.1`;
-    const email = '123@email.com';
-
-    const successResp = {
-      success: true,
-      challenge: {
-        challenge_type: 'email_auth',
-      },
-    };
-
-    const expectedError = new NotificationError({
-      title: 'Auth error',
-      text: 'Invalid or missing email address. Please, try again',
-      type: 'is-danger',
-    });
-
-    it('should make correct request', async () => {
-      expect.assertions(2);
-
-      axiosMock.onPost(url).reply(config => {
-        expect(config.url).toBe(url);
-        expect(config.data).toBe(JSON.stringify({ email }));
-
-        return [200, successResp];
-      });
-
-      await userService.login({ email, redirectUri });
-    });
-
-    it('should handle successfull POST /auth request', async () => {
-      expect.assertions(1);
-
-      axiosMock.onPost(url).reply(200, successResp);
-
-      const challengeType = await userService.login({ email, redirectUri });
-
-      expect(challengeType).toBe(successResp.challenge.challenge_type);
-    });
-
-    it('should handle failed POST /auth request', async () => {
-      axiosMock.onPost(url).reply(200, { success: false });
-      try {
-        await userService.login(email);
-      } catch (receivedError) {
-        expect(receivedError).toEqual(expectedError);
-      }
-    });
-
-    it('should handle rejected POST /auth request', async () => {
-      axiosMock.onPost(url).reply(500, {});
-      try {
-        await userService.login(email);
-      } catch (receivedError) {
-        expect(receivedError).toEqual(expectedError);
-      }
-    });
-  });
-
-  describe('loginViaOTP', () => {
-    const code = 'code';
-    const email = '123@email.com';
-    const url = `${ENV.identityAPIUrl}/token`;
-    const successResp = {
-      success: true,
-    };
-    const expectedError = new NotificationError({
-      title: 'Auth error',
-      text: 'Invalid or missing one time password. Please, try again',
-      type: 'is-danger',
-    });
-
-    it('should make correct request', async () => {
-      axiosMock.onPost(url).reply(config => {
-        expect(config.method).toBe('post');
-        expect(config.url).toBe(url);
-        expect(config.data).toBe(
-          JSON.stringify({
-            challenge_type: 'otp',
-            code,
-            email,
-          }),
-        );
-        return [200, successResp];
-      });
-      await userService.loginViaOTP(code, email);
-    });
-
-    it('should handle successfull POST /token request', async () => {
-      axiosMock.onPost(url).reply(200, successResp);
-      const resp = await userService.loginViaOTP(code, email);
-      expect(resp).toEqual(successResp);
-    });
-
-    it('should handle failed POST /token request', async () => {
-      axiosMock.onPost(url).reply(200, { success: false });
-      try {
-        await userService.loginViaOTP(code, email);
-      } catch (receivedError) {
-        expect(receivedError).toEqual(expectedError);
-      }
-    });
-
-    it('should handle rejected POST /token request', async () => {
-      axiosMock.onPost(url).reply(500, { success: false });
-      try {
-        await userService.loginViaOTP(code, email);
-      } catch (receivedError) {
-        expect(receivedError).toEqual(expectedError);
-      }
-    });
-  });
-
-  describe('logout', () => {
-    const url = `${ENV.identityAPIUrl}/logout`;
-    const expectedError = new NotificationError({
-      title: 'Log out error',
-      text: 'Failed to log out. Please, try again',
-      type: 'is-danger',
-    });
-
-    it('should make correct request', async () => {
-      expect.assertions(1);
-
-      axiosMock.onPost(url).reply(config => {
-        expect(config.url).toBe(url);
-        return [200, {}];
-      });
-
-      await userService.logout();
-    });
-
-    it('should handle rejected POST /logout request', async () => {
-      axiosMock.onPost(url).reply(500, {});
-      try {
-        await userService.logout();
-      } catch (receivedError) {
-        expect(receivedError).toEqual(expectedError);
-      }
-    });
   });
 
   describe('getSettings', () => {
@@ -540,160 +391,6 @@ describe('User service', () => {
       } catch (receivedError) {
         expect(receivedError.text).toEqual(expectedError.text);
       }
-    });
-  });
-
-  describe('Identity mode', () => {
-    const identityModeKey = 'identityMode';
-
-    describe('setIdentityMode', () => {
-      const url = ENV.identityAPIUrl;
-      const type = IDENTITY_MODE.CUSTOM;
-      const mode = { type, serverUrl: url };
-
-      afterEach(() => {
-        localStorage.setItem.mockReset();
-      });
-
-      it('should set the identity mode', () => {
-        expect.assertions(2);
-
-        const spyProxyRequest = jest.spyOn(proxyRequest, 'setMode');
-
-        userService.setIdentityMode(type, url);
-
-        expect(spyProxyRequest).toHaveBeenCalledTimes(1);
-        expect(spyProxyRequest).toBeCalledWith(type, url);
-
-        spyProxyRequest.mockRestore();
-      });
-
-      it('should save the identity mode in the local storage', () => {
-        expect.assertions(2);
-
-        userService.setIdentityMode(type, url);
-
-        const expected = JSON.stringify(mode);
-
-        expect(localStorage.setItem).toHaveBeenCalledTimes(1);
-        expect(localStorage.setItem).toBeCalledWith(identityModeKey, expected);
-      });
-    });
-
-    describe('getIdentityMode', () => {
-      afterEach(() => {
-        localStorage.getItem.mockReset();
-      });
-
-      it('should get the identity mode', () => {
-        userService.getIdentityMode();
-
-        expect(localStorage.getItem).toHaveBeenCalledTimes(1);
-        expect(localStorage.getItem).toBeCalledWith(identityModeKey);
-      });
-
-      it('should return default identity mode', () => {
-        const mode = userService.getIdentityMode();
-
-        expect(mode).toEqual({ type: IDENTITY_MODE.DEFAULT });
-      });
-    });
-
-    describe('deleteIdentityData', () => {
-      beforeEach(() => {
-        proxyRequest.clear = jest.fn();
-      });
-
-      it('should delete identity mode data', async () => {
-        expect.assertions(1);
-
-        await userService.deleteIdentityData();
-
-        expect(proxyRequest.clear).toHaveBeenCalledTimes(1);
-      });
-
-      it('should handle errors', async () => {
-        expect.assertions(1);
-
-        proxyRequest.clear.mockRejectedValueOnce(new Error());
-
-        try {
-          await userService.deleteIdentityData();
-        } catch (e) {
-          expect(e).toBeInstanceOf(NotificationError);
-        }
-      });
-    });
-
-    describe('validateIdentityServer', () => {
-      const serverUrl = 'http://server.com';
-      const url = `${serverUrl}/accounts`;
-      const successResp = [...addresses];
-      const invalidResp = 'invalid response';
-
-      beforeEach(() => {
-        axiosMock = new MockAdapter(axios);
-      });
-
-      it('should make correct request', async () => {
-        expect.assertions(1);
-
-        axiosMock.onGet(url).reply(config => {
-          expect(config.url).toBe(url);
-          return [200, successResp];
-        });
-
-        await userService.validateIdentityServer(serverUrl);
-      });
-
-      it('should return true if the valid custom server', async () => {
-        expect.assertions(1);
-
-        axiosMock.onGet(url).reply(200, successResp);
-
-        const isValid = await userService.validateIdentityServer(serverUrl);
-
-        expect(isValid).toBe(true);
-      });
-
-      it('should throw an error when the response format is invalid', async () => {
-        expect.assertions(2);
-
-        axiosMock.onGet(url).reply(200, invalidResp);
-
-        try {
-          await userService.validateIdentityServer(serverUrl);
-        } catch (e) {
-          expect(e).toBeInstanceOf(NotificationError);
-          expect(e.title).toBe('No Accounts');
-        }
-      });
-
-      it('should throw an error when the response status is 401', async () => {
-        expect.assertions(2);
-
-        axiosMock.onGet(url).reply(401);
-
-        try {
-          await userService.validateIdentityServer(serverUrl);
-        } catch (e) {
-          expect(e).toBeInstanceOf(NotificationError);
-          expect(e.title).toBe('Not Logged In');
-        }
-      });
-
-      it('should throw an error when the response status is invalid', async () => {
-        expect.assertions(2);
-
-        axiosMock.onGet(url).reply(404);
-
-        try {
-          await userService.validateIdentityServer(serverUrl);
-        } catch (e) {
-          expect(e).toBeInstanceOf(NotificationError);
-          expect(e.title).toBe('Invalid Identity Server');
-        }
-      });
     });
   });
 });
